refactor(cli): hoist resolveProjectRef out of the functions command tree
`resolveProjectRef` resolved a project-scoped command's target from an
explicit `--project-ref` or the linked project, and lived in
`next/commands/functions/functions.shared.ts`. It only ever depended on
`ProjectLinkState`, so nothing about it was specific to Edge Functions, and
a second command family needing it would have to reach across command trees
— which `code-structure.unit.test.ts` forbids outright.

Move it to `next/config/`, beside the service it reads, and repoint the
three `functions` handlers. `functions.shared.ts` had no other export, so it
goes rather than being left as a one-line re-export.

Pure move: no behaviour change, and the functions integration tests are
untouched and still pass.
